pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "forge-workflow"
version = "0.1.0"
description = "Portable bot fleet and delivery workflow toolkit"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
]

[project.scripts]
forge = "forge_workflow.cli.main:app"

[tool.setuptools.packages.find]
include = ["forge_workflow*"]

[tool.setuptools.package-data]
forge_workflow = ["templates/**/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=forge_workflow --cov-report=term-missing --cov-report=json:tmp/coverage.json --cov-branch"

[tool.coverage.run]
source = ["forge_workflow"]
branch = true

[tool.coverage.report]
show_missing = true
precision = 1

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
